requirements.txt
pandas~=2.3.3
scikit-learn~=1.7.2
pyyaml~=6.0.3
joblib~=1.5.2
fastapi~=0.119.0
uvicorn

utils~=1.0.2
numpy~=2.3.4
pydantic~=2.12.3
matplotlib~=3.10.7